+++ R/weight_chemo.R
@@ -6,9 +6,13 @@
 #' @param weightSpectral Weight for the spectral score
 #' @param weightBiological Weight for the biological score
 #' @param weightChemical Weight for the chemical consistency score
-#' @param scoreChemicalPathway Score for a `pathway` match (should be lower than `superclass`)
-#' @param scoreChemicalSuperclass Score for a `superclass` match (should be lower than `class`)
-#' @param scoreChemicalClass Score for a `class` match (should be the highest)
+#' @param scoreChemicalClaKingdom Score for a `Classyfire kingdom` match (should be lower than ` Classyfire superclass`)
+#' @param scoreChemicalClaSuperclass Score for a `Classyfire superclass` match (should be lower than `Classyfire class`)
+#' @param scoreChemicalClaClass Score for a `Classyfire class` match (should be lower than `Classyfire parent`)
+#' @param scoreChemicalClaParent Score for a `Classyfire parent` match (should be the highest)
+#' @param scoreChemicalNpcPathway Score for a `pathway` match (should be lower than `superclass`)
+#' @param scoreChemicalNpcSuperclass Score for a `superclass` match (should be lower than `class`)
+#' @param scoreChemicalNpcClass Score for a `class` match (should be the highest)
 #'
 #' @return A table containing the chemically weighted annotation
 #'
@@ -20,9 +24,13 @@ weight_chemo <-
            weightSpectral = weight_spectral,
            weightBiological = weight_biological,
            weightChemical = weight_chemical,
-           scoreChemicalPathway = score_chemical_pathway,
-           scoreChemicalSuperclass = score_chemical_superclass,
-           scoreChemicalClass = score_chemical_class) {
+           scoreChemicalClaKingdom = score_chemical_cla_kingdom,
+           scoreChemicalClaSuperclass = score_chemical_cla_superclass,
+           scoreChemicalClaClass = score_chemical_cla_class,
+           scoreChemicalClaParent = score_chemical_cla_parent,
+           scoreChemicalNpcPathway = score_chemical_npc_pathway,
+           scoreChemicalNpcSuperclass = score_chemical_npc_superclass,
+           scoreChemicalNpcClass = score_chemical_npc_class) {
     log_debug("calculating chemical score ... \n")
     df1 <- annotationTableWeightedBioCleaned
 
@@ -32,41 +40,89 @@ weight_chemo <-
         feature_id,
         structure_inchikey_2D,
         structure_smiles_2D,
-        candidate_structure_1_pathway,
-        candidate_structure_2_superclass,
-        candidate_structure_3_class,
-        consensus_structure_pat,
-        consistency_score_chemical_1_pathway,
-        consensus_structure_sup,
-        consistency_score_chemical_2_superclass,
-        consensus_structure_cla,
-        consistency_score_chemical_3_class
+        candidate_structure_1_cla_kingdom,
+        candidate_structure_1_npc_pathway,
+        candidate_structure_2_cla_superclass,
+        candidate_structure_2_npc_superclass,
+        candidate_structure_3_cla_class,
+        candidate_structure_3_npc_class,
+        candidate_structure_4_cla_parent,
+        consensus_structure_cla_kin,
+        consistency_score_chemical_1_cla_kingdom,
+        consensus_structure_npc_pat,
+        consistency_score_chemical_1_npc_pathway,
+        consensus_structure_cla_sup,
+        consistency_score_chemical_2_cla_superclass,
+        consensus_structure_npc_sup,
+        consistency_score_chemical_2_npc_superclass,
+        consensus_structure_cla_cla,
+        consistency_score_chemical_3_cla_class,
+        consensus_structure_npc_cla,
+        consistency_score_chemical_3_npc_class,
+        consensus_structure_cla_par,
+        consistency_score_chemical_4_cla_parent
       )
 
-    log_debug("... pathway \n")
-    step_pat <- df2 |>
+    log_debug("... (classyfire) kingdom \n")
+    step_cla_kin <- df2 |>
       dplyr::filter(
-        stringr::str_detect(pattern = candidate_structure_1_pathway, string = consensus_structure_pat)
+        stringr::str_detect(pattern = candidate_structure_1_cla_kingdom, string = consensus_structure_cla_kin)
       ) |>
-      dplyr::mutate(score_chemical = scoreChemicalPathway)
+      dplyr::mutate(score_chemical = scoreChemicalClaKingdom)
 
-    log_debug("... superclass \n")
-    step_sup <- step_pat |>
+    log_debug("... (NPC) pathway \n")
+    step_npc_pat <- df2 |>
       dplyr::filter(
-        stringr::str_detect(pattern = candidate_structure_2_superclass, string = consensus_structure_sup)
+        stringr::str_detect(pattern = candidate_structure_1_npc_pathway, string = consensus_structure_npc_pat)
       ) |>
-      dplyr::mutate(score_chemical = scoreChemicalSuperclass)
+      dplyr::mutate(score_chemical = scoreChemicalNpcPathway)
 
-    log_debug("... class \n")
-    step_cla <- step_sup |>
+    log_debug("... (classyfire) superclass \n")
+    step_cla_sup <- step_cla_kin |>
       dplyr::filter(
-        stringr::str_detect(pattern = candidate_structure_3_class, string = consensus_structure_cla)
+        stringr::str_detect(pattern = candidate_structure_2_cla_superclass, string = consensus_structure_cla_sup)
       ) |>
-      dplyr::mutate(score_chemical = scoreChemicalClass)
+      dplyr::mutate(score_chemical = scoreChemicalClaSuperclass)
+
+    log_debug("... (NPC) superclass \n")
+    step_npc_sup <- step_npc_pat |>
+      dplyr::filter(
+        stringr::str_detect(pattern = candidate_structure_2_npc_superclass, string = consensus_structure_npc_sup)
+      ) |>
+      dplyr::mutate(score_chemical = scoreChemicalNpcSuperclass)
+
+    log_debug("... (classyfire) class \n")
+    step_cla_cla <- step_cla_sup |>
+      dplyr::filter(
+        stringr::str_detect(pattern = candidate_structure_3_cla_class, string = consensus_structure_cla_cla)
+      ) |>
+      dplyr::mutate(score_chemical = scoreChemicalClaClass)
+
+    log_debug("... (NPC) class \n")
+    step_npc_cla <- step_npc_sup |>
+      dplyr::filter(
+        stringr::str_detect(pattern = candidate_structure_3_npc_class, string = consensus_structure_npc_cla)
+      ) |>
+      dplyr::mutate(score_chemical = scoreChemicalNpcClass)
+
+    log_debug("... (classyfire) parent \n")
+    step_cla_par <- step_cla_cla |>
+      dplyr::filter(
+        stringr::str_detect(pattern = candidate_structure_4_cla_parent, string = consensus_structure_cla_par)
+      ) |>
+      dplyr::mutate(score_chemical = scoreChemicalClaParent)
 
     log_debug("... outputting best score \n")
     df3 <-
-      dplyr::bind_rows(step_pat, step_sup, step_cla) |>
+      dplyr::bind_rows(
+        step_cla_kin,
+        step_npc_pat,
+        step_cla_sup,
+        step_npc_sup,
+        step_cla_cla,
+        step_npc_cla,
+        step_cla_par
+      ) |>
       dplyr::mutate(dplyr::across(dplyr::where(is.logical), as.numeric)) |>
       dplyr::mutate(score_chemical = ifelse(
         test = is.na(score_chemical),